FIX Remove NullHandler logger override. DebugBar will only display debug and info messages.
Framework will handle what it decides to handle with the HTTP output handler, and whatever is not handled by it will be sent to the DebugBar messages tab. This means the messages tab is unlikely to ever see notices or higher (warnings/errors), but that this module does not interfere with core error handling.